:truck: Move data to version-specific directory
