+++ src/data/JE1.15/nbt_schema.ts
