+++ src/data/JE1.15/registries.ts
